refactor: Move CommandConfigs to its own directory
