+++ packages/melos/test/commands/bootstrap_test.dart
@@ -1,6 +1,7 @@
 import 'dart:io' as io;
 
 import 'package:melos/melos.dart';
+import 'package:melos/src/command_configs/command_configs.dart';
 import 'package:melos/src/commands/runner.dart';
 import 'package:melos/src/common/glob.dart';
 import 'package:melos/src/common/utils.dart';
